tests with Guido's code passed
Former-commit-id: b7e2f5bb287c99c9ab83e7550b43ea8728d754b6

+++ CAnEn/src/Functions.cpp
@@ -112,7 +112,7 @@ Functions::sdCircular(const vector<double> & degs) {
 }
 
 double
-Functions::mean(const vector<double> & values, double max_nan_allowed) {
+Functions::mean(const vector<double> & values, size_t max_nan_allowed) {
     double sum = 0.0;
     size_t nan_count = 0, vec_size = values.size();
